+++ src/main/java/com/guide/c1/BlockDemo.java
@@ -1,4 +1,4 @@
-package com.rossinsky.guide.chapter1;
+package com.guide.c1;
 
 class BlockDemo {
     public static void main(String[] args) {
