+++ src/main/java/com/guide/c1/Example2.java
@@ -1,4 +1,4 @@
-package com.rossinsky.guide.chapter1;
+package com.guide.c1;
 
 class Example2 {
     public static void main(String[] args) {
