+++ src/main/java/com/guide/c1/SelfTest10.java
@@ -1,4 +1,4 @@
-package com.rossinsky.guide.chapter1;
+package com.guide.c1;
 
 /**
  * Prints a conversion table of inches to meters.
