+++ src/main/java/com/guide/c1/SelfTest9.java
@@ -1,4 +1,4 @@
-package com.rossinsky.guide.chapter1;
+package com.guide.c1;
 
 /**
  * Calculates effective weight on the Moon of a man of 80kg.
